完善用户增删

+++ src/main/java/com/tanli/cloud/service/UserManageServiceImp.java
@@ -1,18 +1,24 @@
 package com.tanli.cloud.service;
 
 import com.tanli.cloud.constant.EnvConst;
-import com.tanli.cloud.dao.RepositoryDao;
-import com.tanli.cloud.dao.UserDao;
-import com.tanli.cloud.dao.UserLogDao;
+import com.tanli.cloud.dao.*;
+import com.tanli.cloud.model.ImageInfo;
+import com.tanli.cloud.model.Project;
+import com.tanli.cloud.model.Template;
 import com.tanli.cloud.model.UserLog;
+import com.tanli.cloud.model.response.HarborProject;
 import com.tanli.cloud.model.response.Repository;
 import com.tanli.cloud.model.response.User;
 import com.tanli.cloud.utils.APIResponse;
 import com.tanli.cloud.utils.UuidUtil;
 import org.joda.time.DateTime;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.*;
+import org.springframework.http.client.support.BasicAuthorizationInterceptor;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 
+import javax.annotation.Resource;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -29,6 +35,12 @@ public class UserManageServiceImp implements UserManageService {
     private UserLogDao userLogDao;
     @Autowired
     private RepositoryDao repositoryDao;
+    @Resource
+    private RestTemplate restTemplate;
+    @Autowired
+    private ImageInfoDao imageInfoDao;
+    @Autowired
+    private TemplateDao templateDao;
 
     private static final org.slf4j.Logger LOGGE = org.slf4j.LoggerFactory.getLogger(UserManageServiceImp.class);
 
@@ -37,13 +49,8 @@ public class UserManageServiceImp implements UserManageService {
         //操作日志
         DateTime now = DateTime.now();
         String nowStr = now.getYear()+"-"+now.getMonthOfYear()+"-"+now.getDayOfMonth()+" "+ now.getHourOfDay() + ":"+now.getMinuteOfHour()+":"+now.getSecondOfMinute();
-        UserLog userLog = new UserLog();
-        userLog.setUuid(UuidUtil.getUUID());
-        userLog.setUsername(user.getUserName());
-        userLog.setResoureType("User");
-        userLog.setOperation("登录");
-        userLog.setIsDeleted("0");
-        userLog.setCreate_time(nowStr);
+        UserLog userLog = new UserLog(UuidUtil.getUUID(),user.getUser_uuid(),
+                user.getUserName(),"User", user.getUser_uuid(), "登录", "0", nowStr);
 
         User login_user = userDao.loginVefiry(user);
         userLog.setUser_id(login_user.getUser_uuid());
@@ -75,35 +82,39 @@ public class UserManageServiceImp implements UserManageService {
         repository.setUser_uuid(user.getUser_uuid());
         repository.setRepo_name(user.getUserName()+"_project");
         repository.setRepo_type("private");
-        repository.setUrl("http"+ EnvConst.harbor_api_ip);
-        //repository.set
+        repository.setUrl("http://"+ EnvConst.harbor_api_ip);
+        repository.setLogin_name(EnvConst.harbor_username);
+        repository.setLogin_psd(EnvConst.harbor_password);
+        repository.setProject_name(repository.getRepo_name());
         repository.setCreate_time(nowStr);
         repository.setUpdate_time(nowStr);
 
         //调用Harbor API创建project
-        //在Kubernetes中为用户创建命名空间
-        //将仓库信息保存到数据库中
-        try {
-            //操作日志
-            UserLog userLog = new UserLog();
-            userLog.setUuid(UuidUtil.getUUID());
-            userLog.setUser_id(user.getUser_uuid());
-            userLog.setUsername(user.getUserName());
-            userLog.setResoureType("User");
-            userLog.setResourceId(user.getUser_uuid());
-            userLog.setOperation("增加用户");
-            userLog.setIsDeleted("0");
-            userLog.setCreate_time(nowStr);
-            userLogDao.addUserLog(userLog);
-            int count = userDao.addUser(user);
-            if(count > 0){
-                return APIResponse.success();
-            } else {
+        Boolean result = addHarborProject(repository.getRepo_name());
+        if(result) {
+            try {
+                //操作日志
+                UserLog userLog = new UserLog();
+                userLog.setUuid(UuidUtil.getUUID());
+                userLog.setUser_id(user.getUser_uuid());
+                userLog.setUsername(user.getUserName());
+                userLog.setResoureType("User");
+                userLog.setResourceId(user.getUser_uuid());
+                userLog.setOperation("增加用户");
+                userLog.setIsDeleted("0");
+                userLog.setCreate_time(nowStr);
+                userLogDao.addUserLog(userLog);
+                int count = userDao.addUser(user);
+                repositoryDao.addRepo(repository);
+                if(count > 0){
+                    return APIResponse.success();
+                } else {
+                    LOGGE.info("[UserManageServiceImp Info]: " + "增加用户失败");
+                }
+            } catch (Exception e) {
                 LOGGE.info("[UserManageServiceImp Info]: " + "增加用户失败");
+                e.printStackTrace();
             }
-        } catch (Exception e) {
-            LOGGE.info("[UserManageServiceImp Info]: " + "增加用户失败");
-            e.printStackTrace();
         }
         return APIResponse.fail("增加用户失败");
     }
@@ -111,24 +122,42 @@ public class UserManageServiceImp implements UserManageService {
     @Override
     public APIResponse deleteByIds(User user, String[] ids) {
         Map<String, Integer> result = new HashMap<>();
+        List<ImageInfo> imageInfos = imageInfoDao.getAllImages();
+        List<Template> templateList = templateDao.getAllTemplate();
+
         int success = 0, fail = 0;
         for(int i = 0 ; i < ids.length; i++){
+            //判断用户是否拥有镜像、模板
+            String id = ids[i];
+            List<Repository> repositories = repositoryDao.getRepoByUserid(id);
+            ImageInfo tempImage = imageInfos.stream()
+                    .filter(imageInfo -> imageInfo.getUser_id().equals(id))
+                    .findFirst().orElse(null);
+            Template template = templateList.stream()
+                    .filter(template1 -> template1.getUuid().equals(id))
+                    .findFirst().orElse(null);
+            if(tempImage!=null || template!=null) {//用户id下有应用不能删除
+                fail += 1;
+                break;
+            }
             try {
+                //删除数据库中数据
+                repositoryDao.deleteRepo(id);
+                //删除Harbor中的项目
+                repositories.stream().forEach(repository -> {
+                    deleteHarborProject(repository.getRepo_name());
+                });
+                //删除用户
+                int count = userDao.deleteById(ids[i]);
                 //操作日志
                 DateTime now = DateTime.now();
                 String nowStr = now.getYear()+"-"+now.getMonthOfYear()+"-"+now.getDayOfMonth()+" "+ now.getHourOfDay() + ":"+now.getMinuteOfHour()+":"+now.getSecondOfMinute();
-                UserLog userLog = new UserLog();
-                userLog.setUuid(UuidUtil.getUUID());
-                userLog.setUser_id(user.getUser_uuid());
-                userLog.setUsername(user.getUserName());
-                userLog.setResoureType("User");
-                userLog.setResourceId(ids[i]);
-                userLog.setOperation("删除用户");
-                userLog.setIsDeleted("0");
-                userLog.setCreate_time(nowStr);
+                UserLog userLog = new UserLog(UuidUtil.getUUID(),
+                        user.getUser_uuid(), user.getUserName(), "User", ids[i],
+                        "删除用户",
+                        "0",
+                        nowStr);
                 userLogDao.addUserLog(userLog);
-
-                int count = userDao.deleteById(ids[i]);
                 if(count > 0){
                     success += 1;
                 } else {
@@ -149,8 +178,73 @@ public class UserManageServiceImp implements UserManageService {
         return APIResponse.success(result);
     }
 
+    /**
+     * 使用restTemplate调用harbor API创建project
+     * @param projectName
+     * @return
+     */
     private Boolean addHarborProject(String projectName) {
         Boolean result = false;
+        String url = EnvConst.harbor_api_prefix + "projects";
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_JSON);
+        try {
+            restTemplate.getInterceptors().add(new BasicAuthorizationInterceptor(EnvConst.harbor_username, EnvConst.harbor_password));
+            Project project = new Project(projectName, 0);
+            HttpEntity<Project> request = new HttpEntity<>(project);
+            LOGGE.info("[UserManageServiceImp Info]: " + "POST "+url);
+//            ResponseEntity<String> response = restTemplate.exchange(url, HttpMethod.POST, request, String.class);
+            ResponseEntity<String> response = restTemplate.postForEntity( url, request, String.class);
+            String code = response.getStatusCode().toString();
+            if(code.equals("201")) {
+                return true;
+            }
+        } catch (Exception e) {
+            LOGGE.info("[UserManageServiceImp Info]: " + "向Harbor中添加project失败");
+            e.printStackTrace();
+        }
         return result;
+    }
+
+    private Boolean deleteHarborProject(String project_name){
+        Boolean result = false;
+        HarborProject harborProject = getHarborProjects(project_name);
+        if(harborProject != null) {
+            try {
+                String url = EnvConst.harbor_api_prefix + "projects/" + harborProject.getProject_id();
+                HttpEntity<String> httpEntity = new HttpEntity<String>("");
+                LOGGE.info("[UserManageServiceImp Info]: " + "DELETE " + url);
+                ResponseEntity responseEntity = restTemplate.exchange(url, HttpMethod.DELETE, httpEntity, String.class);
+                if(responseEntity.getStatusCode().toString().equals("200")) {
+                    return true;
+                }
+            }catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+        return result;
+    }
+
+    /**
+     * 查询harbor项目
+     * @param project_name
+     * @return
+     */
+    private HarborProject getHarborProjects(String project_name){
+        HarborProject[] harborProjects = new HarborProject[0];
+        String url = EnvConst.harbor_api_prefix + "projects?project_name=" + project_name;
+        restTemplate.getInterceptors().add(new BasicAuthorizationInterceptor(EnvConst.harbor_username, EnvConst.harbor_password));
+        try {
+            LOGGE.info("[UserManageServiceImp Info]: " + "GET " + url);
+            harborProjects = restTemplate.getForObject(url, HarborProject[].class);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        if(harborProjects.length > 0) {
+            System.out.print(harborProjects[0].toString());
+            return harborProjects[0];
+        } else {
+            return null;
+        }
     }
 }
